linked selection of manager to database

+++ src/main/java/com/admiral/tsp/timesheetportal/data/UserRepository.java
@@ -12,7 +12,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
     User findByUsername(String username);
 
     @Query("SELECT u FROM User u WHERE u.id =:id")
-    User getUserById(@Param("id") Integer id);
+    User getUserById(@Param("id") Long id);
     @Query("SELECT u FROM User u WHERE u.username =:username")
     User getUserByUsername(@Param("username") String username);
     @Query("select u from User u, UserRole a where a.role = 'ROLE_MANAGER' AND u.id = a.userid")
